+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fix: Fix adding integrations on web (#450)
 * Fix: Use `log()` instead of `print()` for SDK logging (#453)
 * Feature: Add `withScope` callback to capture methods (#463)
+* Fix: Add missing `language`, `screenHeightPixels` and `screenWidthPixels` to `SentryDevice`
 
 # 5.0.0